repository 層の実装とテスト追加

+++ backend/repository/pgsql/sqlcgen/query.sql
@@ -3,21 +3,23 @@ select * from users where email = $1 limit 1;
 
 -- name: ListUsers :many
 select
-  id, email, display_name, created_at
+  id, role, email, display_name, created_at
 from users;
 
--- name: InsertUser :exec
+-- name: InsertUser :one
 insert into users (
-  email, display_name, passwd_hash, created_at
-) values ($1, $2, $3, $4);
+  email, role, display_name, passwd_hash, created_at
+) values ($1, $2, $3, $4, $5)
+returning id;
 
 -- name: GetTask :one
-select * from tasks where id = $1 limit 1;
+select * from tasks where id = $1 and user_id=$2 limit 1;
 
--- name: InsertTask :exec
+-- name: InsertTask :one
 insert into tasks (
    user_id, title, created_at
-) values ($1, $2, $3);
+) values ($1, $2, $3)
+returning *;
 
--- name: DeleteTask :exec
-delete from tasks where id = $1;
+-- name: DeleteTask :one
+delete from tasks where id = $1 and user_id=$2 returning *;
